requirements.txt
asgiref==3.5.2
certifi==2022.6.15.1
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.1
django-environ==0.9.0
django-filter==22.1
django-hosts==5.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
Janome==0.4.2
networkx==2.8.7
Pillow==9.2.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
wincertstore==0.2
